# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NoteFlow API",
    "version": "1.0.0",
    "description": "API documentation for the NoteFlow note-taking web application, featuring Role-Based Access Control, versioned routing, and notes CRUD."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Version 1 API Endpoint"
    }
  ],
  "paths": {
...
    "/notes": {
      "get": {
        "summary": "Get notes list",
        "description": "Retrieve list of notes. Users get their own notes. Admins get all notes in the database.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Optional search term for filtering notes by title or content",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": { "type": "string" },
                      "title": { "type": "string" },
                      "content": { "type": "string" },
                      "userId": { "type": "string" },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "updatedAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create a new note",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": { "type": "string", "example": "My First Note" },
                  "content": { "type": "string", "example": "This is the content of my first note." }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": { "type": "string" },
                    "title": { "type": "string" },
                    "content": { "type": "string" },
                    "userId": { "type": "string" },
                    "createdAt": { "type": "string" },
                    "updatedAt": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}